packages/core/tests/rules/unit/expect.test.ts
import { Context, TextSource, ExpectUnitPattern } from '../../../src/index';

/**
 * It can consume the characters 'a', 'b' and 'c' at once respecting the order.
 */
const pattern = new ExpectUnitPattern('a', 'b', 'c');

test('Consume success', () => {
  const context = new Context('test');
  const source = new TextSource('abcabc', context);

  // Test the first consumption.
  expect(pattern.consume(source)).toBeTruthy();
  expect(source.offset).toBe(3);
  expect(source.length).toBe(3);

  // Test the final consumption.
  expect(pattern.consume(source)).toBeTruthy();
  expect(source.offset).toBe(6);
  expect(source.length).toBe(0);
});

test('Consume failure', () => {
  const context = new Context('test');
  const source = new TextSource('azc', context);

  // Test the consumption.
  expect(pattern.consume(source)).toBeFalsy();
  expect(source.offset).toBe(1);
  expect(source.length).toBe(2);
});

test('Consume eof', () => {
  const context = new Context('test');
  const source = new TextSource('a', context);

  // Test the consumption.
  expect(pattern.consume(source)).toBeFalsy();
  expect(source.offset).toBe(1);
  expect(source.length).toBe(0);
});
